Read countries

+++ src/main/java/application/persistence/repository/CountryRepository.java
@@ -3,6 +3,6 @@ package application.persistence.repository;
 import application.persistence.entity.Country;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface CountryRepository extends PagingAndSortingRepository<Country, Long> {
+public interface CountryRepository extends PagingAndSortingRepository<Country, Integer> {
 
 }
